+++ input/drivers_hid/btstack_hid.c
@@ -1439,17 +1439,14 @@ static void btstack_hid_free(const void *data)
       free(hid);
 }
 
-static void *btstack_hid_init(joypad_connection_t *connections)
+static void *btstack_hid_init(void)
 {
    btstack_hid_t *hid = (btstack_hid_t*)calloc(1, sizeof(btstack_hid_t));
 
    if (!hid)
       goto error;
 
-   if(connections == NULL)
-      connections = pad_connection_init(MAX_USERS);
-
-   hid->slots = connections;
+   hid->slots = pad_connection_init(MAX_USERS);
 
    if (!hid->slots)
       goto error;
