+++ input/drivers_hid/iohidmanager_hid.c
@@ -854,7 +854,7 @@ static int iohidmanager_hid_manager_set_device_matching(
    return 0;
 }
 
-static void *iohidmanager_hid_init(joypad_connection_t *connections)
+static void *iohidmanager_hid_init(void)
 {
    iohidmanager_hid_t *hid_apple = (iohidmanager_hid_t*)
       calloc(1, sizeof(*hid_apple));
@@ -862,10 +862,7 @@ static void *iohidmanager_hid_init(joypad_connection_t *connections)
    if (!hid_apple)
       goto error;
 
-   if (connections == NULL)
-      connections = pad_connection_init(MAX_USERS);
-
-   hid_apple->slots = connections;
+   hid_apple->slots = pad_connection_init(MAX_USERS);
 
    if (!hid_apple->slots)
       goto error;
