+++ input/drivers_hid/null_hid.c
@@ -76,10 +76,8 @@ static int16_t null_hid_joypad_axis(void *data, unsigned port, uint32_t joyaxis)
    return 0;
 }
 
-static void *null_hid_init(hid_driver_instance_t *instance)
+static void *null_hid_init(void)
 {
-   (void)instance;
-
    return (null_hid_t*)calloc(1, sizeof(null_hid_t));
 }
