+++ wiiu/include/wiiu/pad_strings.h
@@ -1,0 +1,32 @@
+/*  RetroArch - A frontend for libretro.
+ *  Copyright (C) 2014-2017 - Ali Bouhlel
+ *  Copyright (C) 2011-2017 - Daniel De Matteis
+ *
+ *  RetroArch is free software: you can redistribute it and/or modify it under the terms
+ *  of the GNU General Public License as published by the Free Software Found-
+ *  ation, either version 3 of the License, or (at your option) any later version.
+ *
+ *  RetroArch is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY;
+ *  without even the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR
+ *  PURPOSE.  See the GNU General Public License for more details.
+ *
+ *  You should have received a copy of the GNU General Public License along with RetroArch.
+ *  If not, see <http://www.gnu.org/licenses/>.
+ */
+
+#ifndef __PAD_DRIVER__H
+#define __PAD_DRIVER__H
+
+/**
+ * These are used to map pad names to controller mappings. You can
+ * change these relatively free-form.
+ */
+
+#define PAD_NAME_WIIU_GAMEPAD "WiiU Gamepad"
+#define PAD_NAME_WIIU_PRO "WiiU Pro Controller"
+#define PAD_NAME_WIIMOTE "Wiimote Controller"
+#define PAD_NAME_NUNCHUK "Wiimote+Nunchuk Controller"
+#define PAD_NAME_CLASSIC "Classic Controller"
+#define PAD_NAME_HID "HID Controller"
+
+#endif /* __PAD_DRIVER__H */
