+++ wiiu/input/wiiu_hid_types.h
@@ -1,0 +1,28 @@
+/*  RetroArch - A frontend for libretro.
+ *  Copyright (C) 2013-2014 - Jason Fetters
+ *  Copyright (C) 2011-2017 - Daniel De Matteis
+ *
+ *  RetroArch is free software: you can redistribute it and/or modify it under the terms
+ *  of the GNU General Public License as published by the Free Software Found-
+ *  ation, either version 3 of the License, or (at your option) any later version.
+ *
+ *  RetroArch is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY;
+ *  without even the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR
+ *  PURPOSE.  See the GNU General Public License for more details.
+ *
+ *  You should have received a copy of the GNU General Public License along with RetroArch.
+ *  If not, see <http://www.gnu.org/licenses/>.
+ */
+
+#ifndef __WIIU_HID_TYPES__H
+#define __WIIU_HID_TYPES__H
+
+typedef struct wiiu_hid wiiu_hid_t;
+typedef struct wiiu_adapter wiiu_adapter_t;
+typedef struct wiiu_attach wiiu_attach_event;
+typedef struct _wiiu_event_list wiiu_event_list;
+typedef struct _wiiu_adapter_list wiiu_adapter_list;
+typedef struct _axis_data axis_data;
+typedef struct _wiiu_pad_functions wiiu_pad_functions_t;
+
+#endif /* __WIIU_HID_TYPES__H */
